refactor: 비밀번호 재설정 도메인 변경
