+++ src/main/java/com/sparcs/teamf/api/emailauth/service/EmailAuthService.java
@@ -6,11 +6,9 @@ import static com.sparcs.teamf.domain.emailauth.Event.RESET_PASSWORD;
 import com.sparcs.teamf.api.auth.dto.OneTimeTokenResponse;
 import com.sparcs.teamf.api.auth.jwt.TokenProvider;
 import com.sparcs.teamf.api.emailauth.exception.EmailRequestRequiredException;
-import com.sparcs.teamf.api.emailauth.exception.UnverifiedEmailException;
 import com.sparcs.teamf.api.emailauth.exception.VerificationCodeMismatchException;
 import com.sparcs.teamf.api.member.exception.DuplicateEmailException;
 import com.sparcs.teamf.api.member.exception.MemberNotFoundException;
-import com.sparcs.teamf.api.signup.exception.PasswordMismatchException;
 import com.sparcs.teamf.domain.emailauth.EmailAuth;
 import com.sparcs.teamf.domain.emailauth.EmailAuthRepository;
 import com.sparcs.teamf.domain.member.Member;
@@ -19,7 +17,6 @@ import java.util.concurrent.ThreadLocalRandom;
 import javax.annotation.Resource;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -35,7 +32,6 @@ public class EmailAuthService {
 
     private final EmailAuthRepository emailAuthRepository;
     private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
     private final TokenProvider tokenProvider;
     private final ThreadLocalRandom random = ThreadLocalRandom.current();
 
@@ -75,15 +71,6 @@ public class EmailAuthService {
         return tokenProvider.createOneTimeToken(member.getId(), email);
     }
 
-    public void resetPassword(String email, String password, String confirmPassword) {
-        if (!password.equals(confirmPassword)) {
-            throw new PasswordMismatchException();
-        }
-        handleUnverifiedEmail(email);
-        Member member = memberRepository.findByEmail(email).orElseThrow(MemberNotFoundException::new);
-        member.updatePassword(passwordEncoder.encode(password));
-    }
-
     private boolean isAlreadyRegistered(String email) {
         return memberRepository.existsByEmail(email);
     }
@@ -95,14 +82,6 @@ public class EmailAuthService {
     private void verifyVerificationCodeMismatch(int inputVerificationCode, int savedVerificationCode) {
         if (inputVerificationCode != savedVerificationCode) {
             throw new VerificationCodeMismatchException();
-        }
-    }
-
-    private void handleUnverifiedEmail(String email) {
-        EmailAuth emailAuth = emailAuthRepository.findFirstByEmailAndEventOrderByCreatedDateDesc(email, RESET_PASSWORD)
-                .orElseThrow(EmailRequestRequiredException::new);
-        if (!emailAuth.getIsAuthenticated()) {
-            throw new UnverifiedEmailException();
         }
     }
 }
